Remove PHP-VCR dependency.
I attempted to use PHP-VCR to record the communication to an elasticsearch server,
so the tests would be able to run without an ES server. Its a big hassle and it doesn't
make it easy to test CRUD (because it doesn't see if the action was executed on the server).

I added a Vagrant box with an elasticsearch server in it, which will be used to test this
package on. I think this will also make it easier to see if a new elasticsearch version is
working correctly with the package.