Run elasticsearch in development mode for test builds
When elasticsearch is bound to localhost it runs in development mode. This mode makes a few startup checks of elasticsearch
less strict (they become warnings), allowing it to boot anyway.

On circleci its servers the value for  is too low and elasticsearch wont start. So we make sure we
are running in dev mode so we can at least run our tests

https://github.com/elastic/elasticsearch/issues/4978#issuecomment-257076124